Refactor the event dispatching a bit
The keyper now maintains a mapping of batch index to BatchState. It
doesn't directly write events to a channel.

Instead it let's the BatchState handle those events. BatchState itself
dispatches them by writing them to dedicated channels. This means the
BatchState's Run method can be written in a linear way and we don't
have to worry that we do block the keyper, that's trying to dispatch
the events to different BatchState.