Periodically persist the shutter app state to disk with gob
We use gob to persist the ShutterApp struct to disk. We still need a
way to cleanup the data structures in order to keep the persisted
state small (and also not consume too much memory).

This comes with the drawback that we need to keep the stored state
compatible with changes to the data structures we may want to do (or
have a way to read the old data structures)