Add timeout handling to the deploy command
When asking for a config with 'deploy getconfig' with a config index
that doesn't exist, ganache prints a
'UnhandledPromiseRejectionWarning' and doesn't seem to answer any
requests anymore.

The deploy command now times out if ganache is in that state.

This is more or less just a demonstration on how we should/could
handle timeouts in the keyper service.