Ensure config index is always increasing
In particular, do not allow the next config having the same index as the
current one. This happens naturally if all keypers submit a vote for the
next config, but only the first `threshols` ones are counted. The rest
is interpreted as voting for the same config to be added once more.