Implement staggered tx sending
Keypers now don't send txs at the same time. Instead, for every batch
a single keyper is selected who is expected to send the tx. If they
don't do that in time, the next keyper takes over, and so on. The main
keyper is changed in a round robin fashion.