Start using puredkg to implement the dkg process
While implementing this, I ran into 'go test' hanging. The problem was
TestDKGInstance trying to read from a channel, while the Run method,
that should have put the values on the channel, errored out.
This is fixed now: The MockMessageSender uses a buffered channel and
we run dkg.Run synchronously. In addition, we use selects with default
cases to immediately stop, when no messages where written to the
channel.

This took a bit longer to implement, because