Ignore syncing errors
When syncing either the main chain or shutter state with the observe
package, ignore errors instead of aborting the program.

Make the SyncToHead operate on a cloned object. We will not break the
internal state of these objects in case we run into an error.

Cloning happens by serializing and deserializing with gob. We may
implement a clone method later that doesn't use gob.