Fix observe.Shutter's serialization of encryption keys
We use gob to store (and clone) Shutter objects. The encryption keys
cannot be serialized with gob, because they contain a Params field,
which itself contains a function pointer. This function pointer is nil
when deserializing via gob. As a result we end up with a segfault,
when trying to send an encrypted message as seen in [1].

Encrypted messages are only being sent during the DKG process and this
bug only triggered after a Shutter object had been deserialized from
disk or cloned with medley's CloneWithGob function. This is the reason
that the bug happened only infrequently, even though the encryption
keys must have been in a broken state.

This adds a new type definition, which implements the
GobEncoder/GobDecoder protocol.

I didn't change the CheckIn event to also use this new type, but may
do so in the future.

[1] https://github.com/brainbot-com/shutter/issues/45