Sync to the last committed block height
Previously we asked for events from the latest block. It looks like
this is the next block, which is not committed yet. We didn't get any
events for this block from shuttermint and therefore failed to produce
the correct observe.Shutter state in all cases.

See https://github.com/brainbot-com/shutter/issues/47 and
https://github.com/brainbot-com/shutter/issues/46