Add batch index argument to ExecutorContract functions
This adds a batch index argument to executeCipherBatch,
executePlainBatch and skipCipherBatch.

This allows us to exit early when the batch index doesn't match
without trying to execute the transactions first and later noticing
the hash mismatch. It should reduce the gas usage for keypers that try
to execute the same batch simultaneously for those that come late.

Also it makes the keypers intention clear from just looking at the
ethereum transaction.