Don't look twice for events in current block
Otherwise we potentially pick up the same event multiple times.

As a side effect, we don't look into the first block at all. But this
should be fine as there aren't any events in the genesis block. At a
later point in time, we probably want to set the first CurrentBlock to
the block in which the contracts have been deployed. Then, it will still
be fine as there aren't any interesting events during deployment.