Wait for good time to send txs
If we send txs towards the end of a batch, it's likely that they will
only be included when the next one starts. As a result, the txs fail.

To fix this, we check that the batch won't change in the next few blocks
before we send a tx. If it will, we wait a second and check again.