Make phase length configurable
For simplicity, we just put it in the local config of each keyper. This
means that if a keyper uses a different setting, they'll likely won't be
able to participate.

At some point, we should use shuttermint to agree on the phase length
and maybe automatically increase them if dkg fails.