Allow sending many execution txs at the same time
Prior, we only ever send a single half step execution tx and waited for
it to be confirmed before sending the next one. This is slow, so
catching up after a long stretch of unexecuted batches takes quite a
long of time. Therefore, we now send multiple at the same time if
necessary (but only up to a maximum).