Forbid accusations if the batch is empty
We can make sure the keyper didn't cheat when the cipher batch is
empty, by checking that the caller provides the 'empty cipher batch
hash' iff the list of transactions is empty.

This will allow us to implement some optimizations: keypers won't have
to generate, collect and store the decryption signatures for empty
batches.