Add failing test case for overly strict require
If the cipher hash is zero, the list of transactions must be
empty. The reverse is not true. If a user adds a non-decryptable
transaction, we end up with a non zero cipher hash, but we still have
an empty list of transactions. This test makes sure we can execute a
batch with an empty list of transactions with a non-zero cipher hash.