Run mypy pre-commit in it's own venv
This uses a wrapper script (ported from Raiden) which installs the
project dependencies (via pip-sync) into the pre-commit provided venv.
This removes the need for mypy to be installed in the context in which
git is executed.
Changes to the requirements are automatically detected.

For details see `tools/pre-commit-wrapper.py`.