fix!: only hash written varint bytes
Breaking!

Before, when calculating the mask for the encrypted blocks,
the whole byte-array of length MaxVarint64 was used in the mask hash,
even when not all bytes were written to.

Now, we only consider the written Varint bytes for the hash,
in order to avoid unnecessary keccak256 operations.

Note however that this will break compatibility with messages that were
encrypted before this fix.