feat!: change uint64 byte encoding scheme
The *Little Endian Base 128 VarInt* (LEB128) encoding we used before always
appended 10bytes to each keccak256 block-hash, even if a small integer was used.
Additionally, we don't need the compressibility and variable length properties
of LEB 128 encoding, since the byte encoding is always done locally from known
integers and not transmitted over the wire.

Thus we now use a scheme that encodes uint64 integers directly as a 8byte big-endian byte array.
In order to optimise hash-sizes, unnecessary zero-bytes are omitted from the array starting
from the least-significant byte.
In effect, we encode uint64 as a variable length byte slice with a maximum length of 8.

If the number 0 is encoded, a slice of length 1 with `0x00` value is returned.