IdentifyVersion in VersionMismatch
In order to give a correct error message, we re-parse the mismatched
version with the legacy aware IdentifyVersion. If the marshalled
EncryptedMessage was created before the introduction of the
VersionIdentifier, this will now correctly report the legacy version
`0x00`, instead of just returning the first byte of the marshalled
message.